+++ docker-containers/microservice/src/local_magellan/require_service.py
@@ -3,7 +3,9 @@ from __future__ import print_function
 import argparse
 import os
 import sys
-import xmlrpclib
+import logging
+
+from armada import hermes
 
 import local_magellan
 
@@ -19,42 +21,60 @@ def parse_args():
 
 
 def add_arguments(parser):
-    parser.add_argument('port',
+    parser.add_argument('port', nargs='?',
                         type=int,
                         help='Bind port. Given microservice will be available at localhost:port.')
-    parser.add_argument('microservice_name',
+    parser.add_argument('microservice_name', nargs='?',
                         help='Name of the microservice.')
     parser.add_argument('--env',
                         help='Environment of the microservice. Default: environment variable $MICROSERVICE_ENV.')
     parser.add_argument('--app_id',
                         help='Application ID of the microservice. Default: environment variable $MICROSERVICE_APP_ID.')
+    parser.add_argument('-c', '--config',
+                        help='Name of file with configuration. It should be located in config directory.')
 
 
-def create_magellan_config(port, microservice_name, env, app_id):
+def create_magellan_config_from_file(file_name):
+    config = hermes.get_config(file_name)
+
+    if config:
+        for microservice_name, configurations in config.items():
+            if isinstance(configurations, list):
+                for configuration in configurations:
+                    configure_single_requirement(microservice_name, **configuration)
+
+            elif isinstance(configurations, dict):
+                configure_single_requirement(microservice_name, **configurations)
+    else:
+        logging.warning('Empty dependency file: {}'.format(file_name))
+
+
+def configure_single_requirement(microservice_name, port, env=None, app_id=None):
     microservice = {'microservice_name': microservice_name}
     if env:
         microservice['env'] = env
+    elif 'MICROSERVICE_ENV' in os.environ:
+        microservice['env'] = os.environ.get('MICROSERVICE_ENV')
     if app_id:
         microservice['app_id'] = app_id
-    magellan_config = {
-        port: microservice
-    }
+    elif 'MICROSERVICE_APP_ID' in os.environ:
+        microservice['app_id'] = os.environ.get('MICROSERVICE_APP_ID')
+    magellan_config = {port: microservice}
     local_magellan.save_magellan_config(magellan_config)
 
 
 def main():
     args = parse_args()
-    env = args.env
-    if env is None:
-        env = os.environ.get('MICROSERVICE_ENV')
-    app_id = args.app_id
-    if app_id is None:
-        app_id = os.environ.get('MICROSERVICE_APP_ID')
-    create_magellan_config(args.port, args.microservice_name, env, app_id)
 
-    supervisor_server = xmlrpclib.Server('http://localhost:9001/RPC2')
-    local_magellan_start_output = supervisor_server.supervisor.startProcessGroup('local_magellan')
-    print_err('local_magellan start: {}'.format(local_magellan_start_output))
+    file_name = args.config
+
+    if not file_name:
+        if not args.microservice_name or not args.port:
+            raise RuntimeError('microservice_name and port are required')
+        else:
+            configure_single_requirement(args.microservice_name, args.port, args.env, args.app_id)
+    else:
+        create_magellan_config_from_file(file_name)
 
 
 if __name__ == '__main__':