+++ CHANGELOG.md
@@ -13,6 +13,9 @@ We do best effort to support docker versions 1.12.0 - 17.10.0 with this release.
 - New command `alogs` available in `microservice` base image. It opens up midnight commander with supervisor logs on the
     left pane and service source on the right pane.
 - Add bash completion for supervisorctl
+- Add possibility to configure service requirements via hermes config. This allows us to require different services in different environments.
+    In base microservice image required_service.py will try to find `service_discovery.json` in config path. More info in [docs](http://armada.sh/docs/advanced_features/service_discovery/)
+
 
 ### Improvements
 - Get rid of ugly `WARNING: REMOTE HOST IDENTIFICATION HAS CHANGED!` warnings showing up in remote `armada ssh/diagnose`
@@ -26,7 +29,6 @@ We do best effort to support docker versions 1.12.0 - 17.10.0 with this release.
 - Fix `build-armada` and `restart-armada` development scripts to work outside of armada development vagrant as well.
 - Fix armada's health-check for validating if armada runtime settings have been restored, to work during development
     of armada.
-
 
 
 ## 2.2.0 (2017-10-19)